Checked if home file is writable
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@2203 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/io/HomeFileRecorder.java
@@ -66,7 +66,8 @@ public class HomeFileRecorder implements HomeRecorder {
     File backupFile = null;    
 
     // Backup existing home file to a temporary file
-    if (homeFile.exists()) {
+    if (homeFile.exists()
+        && homeFile.canWrite()) {
       try {
         backupFile = File.createTempFile("backup", ".sh3d");
         copyFile(homeFile, backupFile);
